requirements.txt
Flask==3.0.3
Pillow==11.0.0
python-dotenv==1.0.1
gunicorn==23.0.0

# Testing dependencies
pytest==7.4.4
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-playwright==0.4.4
playwright==1.40.0